Put the files in their final places
